refactor: 사용할 가격 getPriceSummary로 분리
